UC: raise log level for some log messages
This raises the level of some local path and other unmanaged cluster
output that we don't want emitted by default for the Docker Desktop
integration. It also raises the default verbosity of commands so even
though these are logging at a higher level, they will still be printed
in a normal run.

This allows the Docker Desktop integration to explicitly run these calls
with a lower verbosity and not include the messages.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>